requirements.txt
pandas
numpy
transformers
mlflow
pydantic
twikit
google_genai
great_expectations
apache-airflow-providers-microsoft-azure
apache-airflow-providers-postgres
psycopg2-binary
python-dotenv
mlflow